Eliminate compiler warnings
